evaluation/requirements.txt
pandas
datasets
ragas
langchain-openai
python-dotenv
tabulate
